projects/network-config-builder/requirements.txt
# Core dependencies
pyyaml>=6.0
jinja2>=3.1.0
click>=8.1.0
python-dotenv>=1.0.0

# Deployment
paramiko>=3.0.0  # SSH for device deployment

# Vendor API clients (optional)
# librouteros>=3.2.0  # MikroTik API
# pyunifi>=2.26       # UniFi Controller API

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
pylint>=3.0.0
mypy>=1.5.0
